Add rough-bounded row buffer.
This is the core buffer used by Setab. It's count, size, and time
bounded, in that order. The bounding is not exact, as it prefers
violating the constraints over dropping data. It's also not exact,
it does not drop on individual rows, unless you configure the block size
to 1.

Cursors hold the blocks they point to live, until they've consumed it,
so, a single slow consumer can balloon memory usage in the pathological
case.